requirements.txt
# Sistema de Cadastro de Consultas Médicas
# Requirements - Dependências Python

# Framework Web
Flask==3.0.3
Werkzeug==3.0.3

# Banco de Dados
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9

# Utilitários
python-dotenv==1.0.1
Faker==30.8.2

# Servidor WSGI (para produção)
gunicorn==23.0.0
